% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bulk_het.R
\name{plot_bulk_het}
\alias{plot_bulk_het}
\title{plot_bulk_het()}
\usage{
plot_bulk_het(bulk_out, plot_type, components_to_plot, stratifier_gene)
}
\arguments{
\item{bulk_out}{A list containing the output of bulk_het().}

\item{plot_type}{A string indicating the type of plot to generate. Options are 'PC scatter' or 'distance violin'.}

\item{components_to_plot}{A numeric vector indicating the PCA components to plot if plot_type = 'PC scatter'.}

\item{stratifier_gene}{A string indicating the stratifier gene to plot. Required if using quantile-based grouping.}
}
\value{
A ggplot object.
}
\description{
Plot output of bulk_het()
}
\examples{
set.seed(123)
out = bulk_het(data = input_mtx, quant = 0.75, stratifiers = rownames(input_mtx)[1:10], cores = 10) 
plot_bulk_het(bulk_out = out, components_to_plot = c(1,2), plot_type = 'PC scatter', stratifier_gene = 'UBE2Q2P3') + ggplot2::scale_color_brewer(palette = "Dark2")

plot_bulk_het(bulk_out = out, plot_type = 'distance violin', stratifier_gene = 'UBE2Q2P3')
}
