% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scATAC_het.R
\name{plot_scATAC_het}
\alias{plot_scATAC_het}
\title{plot_scATAC_het()}
\usage{
plot_scATAC_het(scATAC_het_out, plot_type = NULL, sample_names = NULL)
}
\arguments{
\item{scATAC_het_out}{A list containing the output of scATAC_het().}

\item{plot_type}{A string indicating the type of plot to generate. Options are 'CV ratios', 'mean ratios', 'CV between samples', or 'mean between samples'.}

\item{sample_names}{A string indicating the sample names to plot, do not include control sample. Default is all samples.}
}
\value{
A ggplot object.
}
\description{
Plot output of scATAC_het()
}
\examples{
plot_scATAC_het(scATAC_het_out, plot_type = 'CV ratios', sample_names = c('RNF8-Ci', 'MIS18A-Ci'))
plot_scATAC_het(scATAC_het_out, plot_type = 'CV between samples', sample_names = c('RNF8-Ci', 'MIS18A-Ci'))
}
