% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_het.R
\name{sc_het}
\alias{sc_het}
\title{sc_het()}
\usage{
sc_het(
  seurat_obj,
  cutoff = 0.1,
  seed = 42,
  sample_cells_per_guide_cutoff = NULL,
  meta_data_sample_column = NULL,
  sample_names = NULL,
  control_sample_name = NULL
)
}
\arguments{
\item{seurat_obj}{A Seurat object containing the scRNA-seq data. The object should be pre-processed for cell quality control only.}

\item{cutoff}{A numeric value indicating the minimum median expression threshold for genes to be included in the analysis. Default is 0.1. It is recommended to adjust this cutoff using the gene_retention_pct() function to retain ~10-20\% of genes.}

\item{seed}{An integer value to set the random seed for reproducibility. Default is 42.}

\item{sample_cells_per_guide_cutoff}{Number of cells to sample per guide. It is highly recommended to choose a number above 50 for a representative outlook of heterogeneity, even at the expense of losing some samples in the analysis. Default uses cell count of the sample with the minimum number of cells.}

\item{meta_data_sample_column}{A string indicating the column name in the Seurat object's metadata that contains the sample information for cells.}

\item{sample_names}{A character vector of sample names to include in the analysis that fall within samples in meta_data_sample_column.}

\item{control_sample_name}{A string indicating the name of the control sample in meta_data_sample_column. This sample will be used as the reference for heterogeneity comparisons.

It is advised that you save the output of this function as an RDS file using saveRDS() immediately after running it for easy loading in future sessions and because memory failure can occur when running downstream plotting functions on large datasets.}
}
\value{
A list with the following components:
\itemize{
\item{guide_subsetted_data}{A list of matrices with the expression data of cells from each sample as well as samples constructed from randomizing cell identity to mitigate issues expected from noisy single cell expression data.}
\item{master_df_list}{A list that contains CV values of each gene within a sample for all samples analyzed.}
\item{mean_shifts_from_NT}{A list that contains gene expression mean of each gene for all samples analyzed.}
\item{asymp_test_p_vals}{A dataframe containing the p-values indicating signiciant change in CV from the cvequality package's asymptotic test for each gene in sample vs control.}
\item{order_of_guides}{A reference for order of samples analyzed for use in plot_sc_het().}
\item{significant_CV_gene_count}{Count of genes in each sample with significant change in CV vs control.}
\item{control_sample_name}{A reference for use in plot_sc_het().}
\item{CV_pvals_adj}{Adjusted p-values for global CV changes of sample vs control using t.test().}
\item{CV_order_of_guides}{A reference for use in plot_sc_het().}
\item{CV_ratios_df}{Ratios of CV values of genes in sample vs control for each sample.}
\item{mean_pvals_adj}{Adjusted p-values for global mean changes of sample vs control using t.test().}
\item{mean_order_of_guides}{A reference for use in plot_sc_het().}
\item{mean_ratios_df}{Ratios of mean values of genes in sample vs control for each sample.}
\item{sample_cells_per_guide_cutoff}{The number of cells sampled per guide for the analysis.}
\item{cc_meta_data}{Metadata of analyzed cells in Seurat object for use in plot_sc_het().}
}
}
\description{
Calculate transcriptional heterogeneity in scRNA-seq samples compared to a control sample.
}
\examples{
sc_het_out = sc_het(seurat_obj = CRISPRa_seurat, cutoff = 0.1, seed = 42, sample_cells_per_guide_cutoff = 50, meta_data_sample_column = 'guide', sample_names = c('NT', 'RNF8', 'MIS18A'), control_sample_name = 'NT')
}
