% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scATAC_het.R
\name{scATAC_het}
\alias{scATAC_het}
\title{scATAC_het()}
\usage{
scATAC_het(
  scATAC_fragment_files = NULL,
  sample_names = NULL,
  control_sample_name = NULL,
  savepath = NULL,
  threads = 10,
  genome = "hg38",
  TSS_bed_path = NULL,
  macs2_path = NULL,
  fixed_cell_count = NULL,
  replicate_map = NULL,
  genes_to_use = NULL,
  samples_to_analyze = NULL,
  seed = 22
)
}
\arguments{
\item{scATAC_fragment_files}{A character vector with file paths to files containing quantified scATAC-seq fragment files (files must end in sample_name.fragments.tsv.gz) using 10x Genomics Cell Ranger ATAC pipeline or other software.}

\item{sample_names}{A character vector with names of the samples corresponding to the fragment files provided.}

\item{control_sample_name}{The name of the control sample.}

\item{savepath}{A file path to a directory where intermediate files and results will be saved.}

\item{threads}{Number of threads to use for ArchR processing. Default is 10.}

\item{genome}{The genome to use for ArchR processing. Default is "hg38".}

\item{TSS_bed_path}{A file path to a bed file containing TSS regions for the genome being used. This is used to create a TSS peak matrix for heterogeneity calculations. Please refer to documentation for ArchR::addPeakSet() for more details.}

\item{macs2_path}{A file path to the MACS2 executable in your current environment. This is used by ArchR to call peaks.}

\item{fixed_cell_count}{An integer specifying a fixed number of cells to use per sample for heterogeneity calculations. If NULL, the minimum number of cells across samples will be used. Default is NULL.}

\item{replicate_map}{A list of length 2 containing vectors for renaming samples if replicates are present. The first vector should contain the original sample names, and the second vector should contain the new names. If NULL, no renaming will be done. Default is NULL.}

\item{genes_to_use}{A character vector of gene names to use for heterogeneity calculations. If NULL, all genes will be used. Default is NULL.}

\item{samples_to_analyze}{A character vector of sample names to include in the heterogeneity calculations. If NULL, all samples will be used. Default is NULL.}

\item{seed}{An integer specifying the seed for random number generation. Default is 22.}
}
\value{
A list containing:
\item{filtered_gene_scores_mtx}{A filtered gene score matrix used for heterogeneity calculations.}
\item{filtered_gene_score_meta_data}{The metadata associated with the filtered gene score matrix.}
\item{control_sample_name}{The name of the control sample.}
\item{CV_ratios_df}{A dataframe containing log2(CV_guide/CV_control) values for each gene across all guides.}
\item{CV_order_of_guides}{The order of guides for CV ratio plotting.}
\item{CV_pvals_adj}{The adjusted p-values from t-tests of log2(CV_guide/CV_control) values for each guide.}
\item{mean_ratios_df}{A dataframe containing log2(mean_guide/mean_control) values for each gene across all guides.}
\item{mean_order_of_guides}{The order of guides for mean ratio plotting.}
\item{mean_pvals_adj}{The adjusted p-values from t-tests of log2(mean_guide/mean_control) values for each guide.}
\item{CVs_df}{A dataframe containing log2(CV_guide) values for each gene across all guides.}
\item{means_df}{A dataframe containing log2(mean_guide) values for each gene across all guides.}
}
\description{
Calculate transcriptional heterogeneity in scATAC-seq samples compared to a control sample.
}
\details{
This function takes in fragment files from scATAC-seq data and calculates transcriptional heterogeneity for each gene in each sample
compared to a control sample. The function uses the ArchR package to process the scATAC-seq data and obtain gene score matrices.
It then calculates the coefficient of variation (CV) for each gene in each sample and compares it to the CV in the control sample.
The function also calculates the mean expression for each gene in each sample and compares it to the mean in the control sample.
The results are returned in a list containing dataframes for CV ratios, mean ratios, and their associated p-values.
}
\examples{
out = scATAC_het(scATAC_fragment_files = paste0(folder, '/', file_names), 
sample_names = names(file_names), 
replicate_map = list(c('NTCi-1', 'NTCi-2', 'NTCa-1', 'NTCa-2', 'RNF8-Ci-1', 'RNF8-Ci-2', 'RNF8-Ca-1', 'RNF8-Ca-2', 'MIS18A-Ci-1', 'MIS18A-Ci-2', 'MIS18A-Ca-1', 'MIS18A-Ca-2'), c('NTCi', 'NTCi', 'NTCa', 'NTCa', 'RNF8-Ci', 'RNF8-Ci', 'RNF8-Ca', 'RNF8-Ca', 'MIS18A-Ci', 'MIS18A-Ci', 'MIS18A-Ca', 'MIS18A-Ca')), ## if replicates present, this replicate map must be done for ALL samples inputted into the function
samples_to_analyze = c('NTCi', 'RNF8-Ci', 'MIS18A-Ci'), control_sample_name = 'NTCi',
savepath = savep, 
threads = 35, 
genome = "hg38", 
TSS_bed_path = '/home/ssobti/projects/heterogeneity_brian/uploaded_data/scATACseq/genome/GRCh38_transcriptsOnly.tss.bed',
macs2_path = '/home/ssobti/miniconda3/envs/archr/bin/macs2') ## please make sure to provide this if using, ArchR gets confused if multiple copies exist
}
